// src/types/buttonEnums.ts
export enum Direction {
  Ascending = "ascending",
  Descending = "descending",
}

export enum Action {
  Add = "add",
  Delete = "delete",
  AddAtHead = "addAtHead",
  AddAtTail = "addAtTail",
  AddAt = "addAt",
  DeleteFromHead = "deleteFromHead",
  DeleteFromTail = "deleteFromTail",
  DeleteFrom = "deleteFrom",
}